refactor: move common fixtures for dataset testing to conftest
